Correct bug in colombia_data
Correction asked in issue #40.

The problem was that in the count of the records were being considered all the records and not just the records with the date after the slice_date. So, after some time, the loop returned an empty data frame. To solve this, I add the parameter `where` in the function used to count the data and add an if to stop the looping while if df_new is empty.